videobox: Fix floating point to integer conversion for the alpha values
Fixes bug #619835.

+++ gst/videobox/gstvideobox.c
@@ -3269,8 +3269,8 @@ static void
 gst_video_box_process (GstVideoBox * video_box, const guint8 * src,
     guint8 * dest)
 {
-  guint b_alpha = CLAMP ((guint) video_box->border_alpha * 256, 0, 256);
-  guint i_alpha = CLAMP ((guint) video_box->alpha * 256, 0, 256);
+  guint b_alpha = CLAMP (video_box->border_alpha * 256, 0, 256);
+  guint i_alpha = CLAMP (video_box->alpha * 256, 0, 256);
   GstVideoBoxFill fill_type = video_box->fill_type;
   gint br, bl, bt, bb, crop_w, crop_h;
